Revise the file search approach
Support for VIIRS Collection 1 and 2 makes it tricky if not
impossible to determine the Level-1 geolocation filename from
the metadata in Level-2. The approach using the Level-2 metadata
to automatically find the associated Level-1 filename and create a
full path to the Level-1 file was replaced by an explicit method
that requires providing a Level-1 filename template.