Milestone: Reproduce the MODIS FRP product
Numerous changes in the classification product and
FRP modules. With these changes the v3.0.x branch
reproduces the MODIS land, cloud and FRP data created
with commit 0d8060 (v2.5.x branch). Water area does
not reproduce, however the relative differences between
the 3.0.x and v2.5.x, gridded at 1x1 degree,
are ~1e-5, which makes me think that this could be
due to round off 'errors' in the data binning caused
by differences (i.e. order of operations/values) in
the aggregation of water and coast pixels. The absolute
value of the differences is less than ~0.5 km2;
differences are positive and negative, so it would
seem that pixels are not double-counted.

To handle the lack of land/water flag associated with
cloud pixels in the VIIRS classification I introduced
a new variable in the output FRP product - area of
unknown surface type pixels. The next step will be to
bring in an auxiliary high resolution land/water mask
to classify the VIIRS cloud pixels as either water or land.